dom/nodes/Node-collection-live.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>NodeList/HTMLCollection liveness Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dom.spec.whatwg.org/#old-style-collections">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>#test5 { display: none }</style>

<div style="display: none">
  <ul id='test1'><li>1</li><li>2</li><li>3</li><li>4</li></ul>
  <ul><li id='test2'>1</li><li>2</li><li>3</li><li>4</li></ul>
  <div id="test3"><b>1</b><b>2</b><b>3</b></div>
  <div id="test4"></div>
  <div id="test5"><p class="test5">1</p><p class="test5">2</p></div>
  <div id="test6"><img src="" alt=""><img src="" alt=""></div>
  <div id="test7"><embed src=""><embed src=""></div>
  <div id="test9"><a href=""></a><a href=""></a></div>
  <div id="test10"><form name=""></form><form name=""></form></div>
  <div id="test12">
    <input name="test12" type="text"/>
    <input name="test12" type="text"/>
  </div>
</div>

<script>

  test(() => {
    var node = document.getElementById("test1");
    var children = node.childNodes;
    var li = document.createElement("li");
    assert_equals(children.length, 4);

    node.appendChild(li);
    assert_equals(children.length, 5);

    node.removeChild(li);
    assert_equals(children.length, 4);
  }, 'Node.childNodes is live');

  test(() => {
    var node = document.getElementById("test2");
    var parentNode = node.parentNode;
    var children = parentNode.children;
    var li = document.createElement("li");
    assert_equals(children.length, 4);

    parentNode.appendChild(li);
    assert_equals(children.length, 5);

    parentNode.removeChild(li);
    assert_equals(children.length, 4);
  }, 'ParentNode.children is live');

  test(() => {
    var element = document.getElementById("test3");
    var elementTag = element.getElementsByTagName("b");
    var documentTag = document.getElementsByTagName("b");
    var b = document.createElement("b");
    assert_equals(elementTag.length, 3);
    assert_equals(documentTag.length, 3);

    element.appendChild(b);
    assert_equals(elementTag.length, 4);
    assert_equals(documentTag.length, 4);

    element.removeChild(b);
    assert_equals(elementTag.length, 3);
    assert_equals(documentTag.length, 3);
  }, 'Document/Element.getElementsByTagName() is live');

  test(() => {
    var element = document.getElementById("test4");
    var span1 = document.createElementNS("foo", "span");
    var span2 = document.createElementNS("foo", "span");
    element.appendChild(span1);
    var elementTag = element.getElementsByTagNameNS("foo", "span");
    var documentTag = document.getElementsByTagNameNS("foo", "span");
    assert_equals(elementTag.length, 1);
    assert_equals(documentTag.length, 1);

    element.appendChild(span2);
    assert_equals(elementTag.length, 2);
    assert_equals(documentTag.length, 2);

    element.removeChild(span1);
    assert_equals(elementTag.length, 1);
    assert_equals(documentTag.length, 1);
  }, 'Document/Element.getElementsByTagNameNS() is live');

  test(() => {
    var element = document.getElementById("test5");
    var elementTag = element.getElementsByClassName("test5");
    var documentTag = document.getElementsByClassName("test5");
    var p = document.createElement("p");
    p.setAttribute("class", "test5");
    assert_equals(elementTag.length, 2);
    assert_equals(documentTag.length, 2);

    element.appendChild(p);
    assert_equals(elementTag.length, 3);
    assert_equals(documentTag.length, 3);

    element.removeChild(p);
    assert_equals(elementTag.length, 2);
    assert_equals(documentTag.length, 2);
  }, 'Document/Element.getElementsByClassName() is live');

  test(() => {
    var element = document.getElementById("test6");
    var imgs = document.images;
    var img = document.createElement("img");
    assert_equals(imgs.length, 2);

    element.appendChild(img);
    assert_equals(imgs.length, 3);

    element.removeChild(img);
    assert_equals(imgs.length, 2);
  }, 'Document.images is live');

  test(() => {
    var element = document.getElementById("test7");
    var embeds = document.embeds;
    var embed = document.createElement("embed");
    assert_equals(embeds.length, 2);

    element.appendChild(embed);
    assert_equals(embeds.length, 3);

    element.removeChild(embed);
    assert_equals(embeds.length, 2);
  }, 'Document.embeds is live');

  test(() => {
    var element = document.getElementById("test7");
    var Plugins = document.plugins;
    var embed = document.createElement("embed");
    assert_equals(Plugins.length, 2);

    element.appendChild(embed);
    assert_equals(Plugins.length, 3);

    element.removeChild(embed);
    assert_equals(Plugins.length, 2);
  }, 'Document.plugins is live');

  test(() => {
    var element = document.getElementById("test9");
    var links = document.links;
    var a = document.createElement("a");
    a.setAttribute("href", "");
    assert_equals(links.length, 2);

    element.appendChild(a);
    assert_equals(links.length, 3);

    element.removeChild(a);
    assert_equals(links.length, 2);
  }, 'Document.links is live');

  test(() => {
    var element = document.getElementById("test10");
    var forms = document.forms;
    var form = document.createElement("form");
    assert_equals(forms.length, 2);

    element.appendChild(form);
    assert_equals(forms.length, 3);

    element.removeChild(form);
    assert_equals(forms.length, 2);
  }, 'Document.forms is live');

  test(() => {
    var scripts = document.scripts;
    var script = document.createElement("script");
    assert_equals(scripts.length, 3);

    document.body.appendChild(script);
    assert_equals(scripts.length, 4);

    document.body.removeChild(script);
    assert_equals(scripts.length, 3);
  }, 'Document.scripts is live');

  test(() => {
    var element = document.getElementById("test12");
    var elements = document.getElementsByName("test12");
    var input = document.createElement("input");
    input.setAttribute("name", "test12");
    input.setAttribute("type", "text");
    assert_equals(elements.length, 2);

    element.appendChild(input);
    assert_equals(elements.length, 3);

    element.removeChild(input);
    assert_equals(elements.length, 2);
  }, 'Document.getElementsByName() is live');

</script>
